+++ src/screens/validator_details/components/nodes/index.tsx
@@ -1,9 +1,12 @@
 import React from 'react';
+import classnames from 'classnames';
 import dynamic from 'next/dynamic';
 import {
   usePagination,
   useScreenSize,
 } from '@hooks';
+import { Typography } from '@material-ui/core';
+import useTranslation from 'next-translate/useTranslation';
 import {
   Pagination,
   NoData,
@@ -19,6 +22,7 @@ const Desktop = dynamic(() => import('./components/desktop'));
 const Mobile = dynamic(() => import('./components/mobile'));
 
 const Nodes: React.FC<ComponentDefault> = (props) => {
+  const { t } = useTranslation('validators');
   const { isDesktop } = useScreenSize();
   const classes = useStyles();
   const {
@@ -47,7 +51,8 @@ const Nodes: React.FC<ComponentDefault> = (props) => {
   }
 
   return (
-    <Box className={props.className}>
+    <Box className={classnames(props.className, classes.root)}>
+      <Typography variant="h2">{t('nodes')}</Typography>
       {component}
       <Pagination
         className={classes.paginate}