+++ src/screens/validator_details/components/overview/index.tsx
@@ -3,12 +3,25 @@ import useTranslation from 'next-translate/useTranslation';
 import {
   BoxDetails,
 } from '@components';
+import { getMiddleEllipsis } from '@utils/get_middle_ellipsis';
 import { OverviewType } from '../../types';
 
 const Overview: React.FC<{overview: OverviewType} & ComponentDefault> = (props) => {
   const { t } = useTranslation('validators');
-  const details = [
-    {
+  const details = [];
+
+  props.overview.stakeDistribution.forEach((x) => {
+    const key = getMiddleEllipsis(x.key, {
+      beginning: 10, ending: 7,
+    });
+    details.push({
+      label: `${t('distribution')} (${key})`,
+      detail: `${x.value}`,
+    });
+  });
+
+  if (props.overview.stakeDistribution.length <= 2) {
+    details.push({
       label: t('website'),
       detail: (
         props.overview.website ? (
@@ -21,16 +34,22 @@ const Overview: React.FC<{overview: OverviewType} & ComponentDefault> = (props) 
           </a>
         ) : '-'
       ),
-    },
-    {
+    });
+  }
+
+  if (props.overview.stakeDistribution.length <= 1) {
+    details.push({
       label: t('location'),
       detail: props.overview.location || '-',
-    },
-    {
+    });
+  }
+
+  if (props.overview.stakeDistribution.length <= 0) {
+    details.push({
       label: t('identity'),
       detail: props.overview.identity || '-',
-    },
-  ];
+    });
+  }
 
   return (
     <BoxDetails