+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - Add Transaction Details Page
 - Add Validators Page
 - Add Node Details Page
+- Add Validator Details Page